+++ src/fsw/fsw-symbol-size.js
@@ -1,5 +1,6 @@
 
 import { key2id } from '@sutton-signwriting/core/convert/convert'; 
+import { parse } from '@sutton-signwriting/core/fsw/fsw';
 import { symbolSize as symSize } from '../font/font-symbol-size';
 
 /**
@@ -13,6 +14,10 @@ import { symbolSize as symSize } from '../font/font-symbol-size';
  * return [15,30]
  */
 const symbolSize = function (fsw){
+  const parsed = parse.symbol(fsw);
+  if (!parsed.symbol) {
+    return undefined;
+  }
   return symSize(key2id(fsw));
 }