+++ src/fsw/index.js
@@ -9,6 +9,10 @@ import { symbolLine, symbolFill, symbolText } from  './fsw-symbol-text';
 import { symbolSvgBody, symbolSvg } from  './fsw-symbol-svg';
 import { symbolPng } from './fsw-symbol-png';
 import { symbolNormalize } from './fsw-symbol-normalize';
+import { symbolMirror } from './fsw-symbol-mirror';
+import { symbolRotate } from './fsw-symbol-rotate';
+import { symbolFlop } from './fsw-symbol-flop';
+import { symbolScroll } from './fsw-symbol-scroll';
 import { signSvgBody, signSvg } from './fsw-sign-svg';
 import { signPng } from './fsw-sign-png';
 import { signNormalize } from './fsw-sign-normalize';
@@ -17,4 +21,4 @@ import { columnsSvg } from './fsw-columns-svg';
 import { columnPng } from './fsw-column-png';
 import { columnsPng } from './fsw-columns-png';
 
-export { symbolSize, symbolLine, symbolFill, symbolText, symbolSvgBody, symbolSvg, symbolPng, symbolNormalize, signSvgBody, signSvg, signPng, signNormalize, columnSvg, columnsSvg, columnPng, columnsPng }
+export { symbolSize, symbolLine, symbolFill, symbolText, symbolSvgBody, symbolSvg, symbolPng, symbolNormalize, symbolMirror, symbolRotate, symbolFlop, symbolScroll, signSvgBody, signSvg, signPng, signNormalize, columnSvg, columnsSvg, columnPng, columnsPng }