+++ src/swu/index.js
@@ -9,6 +9,10 @@ import { symbolLine, symbolFill, symbolText } from  './swu-symbol-text';
 import { symbolSvgBody, symbolSvg } from  './swu-symbol-svg';
 import { symbolPng } from './swu-symbol-png';
 import { symbolNormalize } from './swu-symbol-normalize';
+import { symbolMirror } from './swu-symbol-mirror';
+import { symbolRotate } from './swu-symbol-rotate';
+import { symbolFlop } from './swu-symbol-flop';
+import { symbolScroll } from './swu-symbol-scroll';
 import { signSvgBody, signSvg } from './swu-sign-svg';
 import { signPng } from './swu-sign-png';
 import { signNormalize } from './swu-sign-normalize';
@@ -17,4 +21,4 @@ import { columnsSvg } from './swu-columns-svg';
 import { columnPng } from './swu-column-png';
 import { columnsPng } from './swu-columns-png';
 
-export { symbolSize, symbolLine, symbolFill, symbolText, symbolSvgBody, symbolSvg, symbolPng, symbolNormalize, signSvgBody, signSvg, signPng, signNormalize, columnSvg, columnsSvg, columnPng, columnsPng }
+export { symbolSize, symbolLine, symbolFill, symbolText, symbolSvgBody, symbolSvg, symbolPng, symbolNormalize, symbolMirror, symbolRotate, symbolFlop, symbolScroll, signSvgBody, signSvg, signPng, signNormalize, columnSvg, columnsSvg, columnPng, columnsPng }