+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Todo
 - ...
 
+## [1.5.0] - 2023-03-11
+### Added
+- symbol transformations: mirror, rotate, flop, and scroll
+
+### Updated
+- tests and documentation
+
+### Fixed
+- symbol size for bad input
+
 ## [1.4.5] - 2023-02-11
 ### Fixed
 - frequent canvas read warning
@@ -79,7 +89,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - documentation for core modules
 
 
-[Unreleased]: https://github.com/sutton-signwriting/font-ttf/compare/v1.4.5...HEAD
+[Unreleased]: https://github.com/sutton-signwriting/font-ttf/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/sutton-signwriting/font-ttf/releases/tag/v1.5.0
 [1.4.5]: https://github.com/sutton-signwriting/font-ttf/releases/tag/v1.4.5
 [1.4.4]: https://github.com/sutton-signwriting/font-ttf/releases/tag/v1.4.4
 [1.4.3]: https://github.com/sutton-signwriting/font-ttf/releases/tag/v1.4.3